+++ public/react-services/saved-objects.js
@@ -11,7 +11,9 @@
  */
 
 import { GenericRequest } from './generic-request';
-import { KnownFields, FieldsStatistics, FieldsMonitoring } from '../utils/known-fields'
+import { KnownFields } from '../utils/known-fields'
+import { FieldsStatistics } from '../utils/statistics-fields'
+import { FieldsMonitoring } from '../utils/monitoring-fields'
 export class SavedObject {
   /**
    *
@@ -79,7 +81,7 @@ export class SavedObject {
   static async existsOrCreateIndexPattern(patternID) {
     const result = await SavedObject.existsIndexPattern(patternID);
     if (!result.data) {
-      const fields = await SavedObject.getIndicesFields(patternID);
+      const fields = await SavedObject.getIndicesFields(patternID,'alerts');
       await this.createSavedObject(
         'index-pattern',
         patternID,
@@ -170,7 +172,7 @@ export class SavedObject {
    */
   static async refreshIndexPattern(pattern) {
     try {
-      const fields = await SavedObject.getIndicesFields(pattern.title);
+      const fields = await SavedObject.getIndicesFields(pattern.title,'alerts');
       await this.refreshFieldsOfIndexPattern(pattern.id, pattern.title, fields);
 
       return;
@@ -187,7 +189,7 @@ export class SavedObject {
    */
   static async createWazuhIndexPattern(pattern) {
     try {
-      const fields = await SavedObject.getIndicesFields(pattern);
+      const fields = await SavedObject.getIndicesFields(pattern,'alerts');
       await this.createSavedObject(
         'index-pattern',
         pattern,
@@ -213,17 +215,17 @@ export class SavedObject {
     }
   }
 
-  static getIndicesFields = async (pattern) => GenericRequest.request(
+  static getIndicesFields = async (pattern,selectedPattern) => GenericRequest.request(
     //we check if indices exist before creating the index pattern
     'GET',
     `/api/index_patterns/_fields_for_wildcard?pattern=${pattern}&meta_fields=_source&meta_fields=_id&meta_fields=_type&meta_fields=_index&meta_fields=_score`,
     {}
   ).then(response => response.data.fields).catch(() => {
-    if (pattern === 'wazuh-monitoring-*') {
+    if (selectedPattern === 'monitoring') {
       return FieldsMonitoring;
-    } else if (pattern === 'wazuh-statistics-*') {
+    } else if (selectedPattern === 'statistics') {
       return FieldsStatistics;
-    } else {
+    } else if (selectedPattern === 'alerts'){
       return KnownFields
     }
   })
