+++ public/utils/known-fields.js
@@ -1,70 +1,3 @@
-export const FieldsStatistics = [
-  { "name": "analysisd.alerts_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.alerts_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.alerts_written", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.archives_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.archives_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.dbsync_mdps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.dbsync_messages_dispatched", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.dbsync_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.dbsync_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.event_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.event_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.events_dropped", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.events_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.events_processed", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.events_received", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.firewall_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.firewall_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.firewall_written", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.fts_written", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.hostinfo_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.hostinfo_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.hostinfo_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.hostinfo_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.other_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.other_events_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.rootcheck_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.rootcheck_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.rootcheck_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.rootcheck_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.rule_matching_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.rule_matching_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.sca_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.sca_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.sca_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.sca_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.statistical_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.statistical_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscheck_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscheck_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscheck_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscheck_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscollector_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscollector_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscollector_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.syscollector_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.total_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.winevt_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.winevt_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.winevt_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "analysisd.winevt_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "apiName", "type": "string", "esTypes": ["text"], "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "apiName.keyword", "type": "string", "esTypes": ["keyword"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true, "subType": { "multi": { "parent": "apiName" } } },
-  { "name": "cluster", "type": "string", "esTypes": ["text"], "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "cluster.keyword", "type": "string", "esTypes": ["keyword"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true, "subType": { "multi": { "parent": "cluster" } } },
-  { "name": "nodeName", "type": "string", "esTypes": ["text"], "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "nodeName.keyword", "type": "string", "esTypes": ["keyword"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true, "subType": { "multi": { "parent": "nodeName" } } },
-  { "name": "remoted.ctrl_msg_count", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.dequeued_after_close", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.discarded_count", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.evt_count", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.msg_sent", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.recv_bytes", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.tcp_sessions", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "remoted.total_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "timestamp", "type": "date", "esTypes": ["date"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true }]
 export const KnownFields = [
   {
     "name": "_id",
@@ -3964,32 +3897,3 @@ export const KnownFields = [
     "readFromDocValues": false
   }
 ];
-export const FieldsMonitoring =
-  [{ "name": "timestamp", "type": "date", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "_id", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": false },
-  { "name": "_index", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": false },
-  { "name": "_score", "type": "number", "count": 0, "scripted": false, "searchable": false, "aggregatable": false, "readFromDocValues": false },
-  { "name": "_source", "type": "_source", "count": 0, "scripted": false, "searchable": false, "aggregatable": false, "readFromDocValues": false },
-  { "name": "_type", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": false },
-  { "name": "dateAdd", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "group", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "host", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "id", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "ip", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "lastKeepAlive", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "cluster.name", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "mergedSum", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "configSum", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "node_name", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "manager", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "manager_host", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "name", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "os.arch", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "os.codename", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "os.major", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "os.name", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "os.platform", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "os.uname", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "os.version", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
-  { "name": "status", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
-  { "name": "version", "type": "string", "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false }]
