+++ public/utils/statistics-fields.ts
@@ -1,0 +1,67 @@
+export const FieldsStatistics = [
+  { "name": "analysisd.alerts_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.alerts_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.alerts_written", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.archives_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.archives_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.dbsync_mdps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.dbsync_messages_dispatched", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.dbsync_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.dbsync_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.event_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.event_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.events_dropped", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.events_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.events_processed", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.events_received", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.firewall_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.firewall_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.firewall_written", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.fts_written", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.hostinfo_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.hostinfo_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.hostinfo_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.hostinfo_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.other_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.other_events_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.rootcheck_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.rootcheck_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.rootcheck_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.rootcheck_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.rule_matching_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.rule_matching_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.sca_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.sca_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.sca_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.sca_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.statistical_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.statistical_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscheck_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscheck_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscheck_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscheck_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscollector_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscollector_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscollector_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.syscollector_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.total_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.winevt_edps", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.winevt_events_decoded", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.winevt_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "analysisd.winevt_queue_usage", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "apiName", "type": "string", "esTypes": ["text"], "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
+  { "name": "apiName.keyword", "type": "string", "esTypes": ["keyword"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true, "subType": { "multi": { "parent": "apiName" } } },
+  { "name": "cluster", "type": "string", "esTypes": ["text"], "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
+  { "name": "cluster.keyword", "type": "string", "esTypes": ["keyword"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true, "subType": { "multi": { "parent": "cluster" } } },
+  { "name": "nodeName", "type": "string", "esTypes": ["text"], "count": 0, "scripted": false, "searchable": true, "aggregatable": false, "readFromDocValues": false },
+  { "name": "nodeName.keyword", "type": "string", "esTypes": ["keyword"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true, "subType": { "multi": { "parent": "nodeName" } } },
+  { "name": "remoted.ctrl_msg_count", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.dequeued_after_close", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.discarded_count", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.evt_count", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.msg_sent", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.recv_bytes", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.tcp_sessions", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "remoted.total_queue_size", "type": "number", "esTypes": ["long"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true },
+  { "name": "timestamp", "type": "date", "esTypes": ["date"], "count": 0, "scripted": false, "searchable": true, "aggregatable": true, "readFromDocValues": true }]
